@@ -1,6 +1,6 @@
 {
   "name": "ngx-input-file",
-  "version": "9.0.0",
+  "version": "10.0.0",
   "description": "Angular Input File Component",
   "license": "MIT",
   "repository": {
@@ -30,49 +30,45 @@
   },
   "private": false,
   "peerDependencies": {
-    "@angular/cdk": "^9.2.4",
-    "@angular/animations": "^9.1.12",
-    "@angular/material": "^9.2.4",
-    "@angular/common": "~9.1.12",
-    "@angular/core": "~9.1.12",
-    "@angular/forms": "~9.1.12"
+    "@angular/cdk": "^10.2.7",
+    "@angular/material": "^9.2.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.12",
-    "@angular-devkit/build-ng-packagr": "~0.901.12",
-    "@angular/animations": "^9.1.12",
-    "@angular/cdk": "^9.2.4",
-    "@angular/cli": "~9.1.12",
-    "@angular/common": "^9.1.12",
-    "@angular/compiler": "~9.1.12",
-    "@angular/compiler-cli": "~9.1.12",
-    "@angular/core": "^9.1.12",
-    "@angular/forms": "^9.1.12",
-    "@angular/language-service": "~9.1.12",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular-devkit/build-ng-packagr": "~0.1002.0",
+    "@angular/animations": "^10.2.3",
+    "@angular/cdk": "^10.2.7",
+    "@angular/cli": "~10.2.0",
+    "@angular/common": "^10.2.3",
+    "@angular/compiler": "~10.2.3",
+    "@angular/compiler-cli": "~10.2.3",
+    "@angular/core": "^10.2.3",
+    "@angular/forms": "^10.2.3",
+    "@angular/language-service": "~10.2.3",
     "@angular/material": "^9.2.4",
-    "@angular/platform-browser": "~9.1.12",
-    "@angular/platform-browser-dynamic": "~9.1.12",
-    "@angular/router": "~9.1.12",
+    "@angular/platform-browser": "~10.2.3",
+    "@angular/platform-browser-dynamic": "~10.2.3",
+    "@angular/router": "~10.2.3",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^12.12.6",
     "bootstrap": "^4.3.1",
     "codelyzer": "^5.1.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~6.0.0",
-    "karma": "~5.2.3",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "material-design-icons": "^3.0.1",
-    "ng-packagr": "~9.1.5",
-    "protractor": "~5.4.2",
+    "ng-packagr": "^10.1.0",
+    "protractor": "~7.0.0",
     "rxjs": "~6.6.3",
     "ts-node": "~8.3.0",
-    "tslib": "^2.0.3",
-    "tslint": "~5.18.0",
-    "typescript": "~3.8.3",
+    "tslib": "^2.0.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.5",
     "zone.js": "~0.10.3"
   }
 }
