@@ -34,7 +34,7 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 })
 
 export class InputFileModule {
-    public static forRoot(config: InputFileConfig): ModuleWithProviders {
+    public static forRoot(config: InputFileConfig): ModuleWithProviders<InputFileModule> {
         return {
             ngModule: InputFileModule,
             providers: [
